backend/requirements.txt
# Smart Scraper Platform - Backend Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1

# Configuration
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<5

# Task Queue
celery[redis]>=5.3.6
redis>=5.0.1

# Web Scraping
playwright>=1.41.0
httpx>=0.26.0
beautifulsoup4>=4.12.3
lxml>=5.2.1

# Vector Database
faiss-cpu>=1.7.4
numpy>=1.26.0

# AI/LLM
langchain>=0.1.0
langgraph>=0.0.20
crewai>=0.1.0
openai>=1.8.0

# Export
openpyxl>=3.1.2
python-docx>=1.1.0
reportlab>=4.0.8

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0
aiosqlite>=0.19.0

# Utilities
python-dotenv>=1.0.0
bleach>=6.1.0
